+++ src/pages/SAMPLE_PAGE.tsx
@@ -2,23 +2,13 @@ import React, { ReactElement } from 'react';
 import { Row, Col } from 'react-bootstrap';
 // import { useDispatch } from 'react-redux';
 
-// *** Import selectors ***
+// eslint-disable-next-line @typescript-eslint/ban-types
+type SamplePageProps = {};
 
-// *** Import reducers ***
-
-// *** Import components ***
-
-// *** Main component ***
-const SamplePage = (): ReactElement => {
+// eslint-disable-next-line no-empty-pattern
+const SamplePage = ({}: SamplePageProps): ReactElement => {
   // const dispatch = useDispatch();
 
-  // *** Selectors ***
-  // const items: SampleItemType[] = useGetItems();
-
-  // *** State (useState) ***
-  //  const [myMessage, setMyMessage] = React.useState('');
-
-  // *** On component load ***
   // eslint-disable-next-line arrow-body-style
   React.useEffect(() => {
     // *** items run on component mount ***
@@ -29,18 +19,11 @@ const SamplePage = (): ReactElement => {
   }, []);
 
   return (
-    <div data-testid='pageContent'>
-      <Row>
-        <Col>
-          {/* Page title */}
-          <h1>Sample Page</h1>
-
-          {/* Page content */}
-
-          {/* End page content */}
-        </Col>
-      </Row>
-    </div>
+    <Row>
+      <Col>
+        <h1>Sample Page</h1>
+      </Col>
+    </Row>
   );
 };
